+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add new `Progress` React component.
+- Fix: Propagate props from `CallToActionPopover` to `Popover` components.
+- Fix: Add key to buttons rendering in `CallToActionPopover`.
 
 ## [5.9.0] - 2017-02-06